move valhalla stuff to new ValhallClient module
