--- web/frontend/src/utils/routecache.ts
@@ -1,71 +1,0 @@
-import { POI, DistanceUnits } from './models';
-import { ProcessedRouteSummary, Route, summarizeRoute } from './routes';
-
-export type CacheableMode = 'walk' | 'bicycle' | 'car';
-
-function modeToCostingModel(mode: CacheableMode): string {
-  switch (mode) {
-    case 'walk':
-      return 'pedestrian';
-    case 'bicycle':
-      return 'bicycle';
-    case 'car':
-      return 'auto';
-  }
-}
-
-export async function getRoutes(
-  from: POI,
-  to: POI,
-  mode: CacheableMode,
-  units?: DistanceUnits
-): Promise<[Route, ProcessedRouteSummary][]> {
-  if (!from.position || !to.position) {
-    console.error("Can't request without fully specified endpoints");
-    return [];
-  }
-
-  type RouteRequest = {
-    locations: Array<{ lat: number; lon: number }>;
-    costing: string;
-    alternates: number;
-    units?: DistanceUnits;
-  };
-  const requestObject: RouteRequest = {
-    locations: [
-      {
-        lat: from.position.lat,
-        lon: from.position.long,
-      },
-      {
-        lat: to.position.lat,
-        lon: to.position.long,
-      },
-    ],
-    costing: modeToCostingModel(mode),
-    alternates: 3,
-  };
-  if (units) {
-    requestObject.units = units;
-  }
-  const response = await fetch(
-    `/valhalla/route?json=${JSON.stringify(requestObject)}`
-  );
-  if (response.status !== 200) {
-    console.error('Valhalla response gave error: ' + response.status);
-    return [];
-  }
-  const responseJson = await response.json();
-  const routes: [Route, ProcessedRouteSummary][] = [];
-  const route = responseJson.trip as Route;
-  if (route) {
-    routes.push([route, summarizeRoute(route)]);
-  }
-  for (const altIdx in responseJson.alternates) {
-    const route = responseJson.alternates[altIdx].trip as Route;
-    if (route) {
-      routes.push([route, summarizeRoute(route)]);
-    }
-  }
-  return routes;
-}
